+++ src/main/java/com/project/skin/category/service/dto/ReadCategory.java
@@ -1,4 +1,4 @@
-package com.project.skin.category.dto;
+package com.project.skin.category.service.dto;
 
 import lombok.Builder;
 import lombok.Getter;
